Move FatalError to internal
