+++ cmd/once/cmdonce/internal/error.go
@@ -1,0 +1,10 @@
+package internal
+
+import (
+	"errors"
+)
+
+var (
+	ErrNoDocker           = errors.New("no docker")
+	ErrDockerVolumeExists = errors.New("docker volume exists")
+)
